TS-40724 Refactorings
